+++ src/dev/Renderer/pages/Templates/Template/index.js
@@ -87,6 +87,7 @@ class Template extends Page {
           status: id ? 200 : 404,
         })
       }}
+      inEditMode={true}
       {...other}
     />
   }